+++ packages/brawl-stars-api/src/lib/api/players.ts
@@ -1,0 +1,23 @@
+import { formatTag } from "@brawltracker/supercell-api-utils";
+import type { $Fetch } from "ofetch";
+import type { BattlelogsResponse } from "../types/battlelogs";
+import type { Player } from "../types/players";
+
+export class PlayersApi {
+	private readonly $fetch: $Fetch;
+
+	public constructor($fetch: $Fetch) {
+		this.$fetch = $fetch;
+	}
+
+	public async getPlayer(tag: string) {
+		return this.$fetch<Player>(`/players/${encodeURIComponent(formatTag(tag))}`);
+	}
+
+	public async getPlayerBattleLog(tag: string) {
+		const { items } = await this.$fetch<BattlelogsResponse>(
+			`/players/${encodeURIComponent(formatTag(tag))}/battlelog`
+		);
+		return items;
+	}
+}