+++ packages/brawl-stars-api/src/lib/types/api.ts
@@ -1,0 +1,4 @@
+export interface ApiJsonError {
+	message?: string;
+	reason: string;
+}