+++ packages/brawl-stars-api/src/lib/types/brawlers.ts
@@ -1,0 +1,19 @@
+import type { StarPower, Gadget } from "./powers";
+
+export interface Brawler {
+	gadgets: Gadget[];
+	id: number;
+	name: string;
+	starPowers: StarPower[];
+}
+
+export interface BasicBrawler {
+	id: number;
+	name: string;
+	power: number;
+	trophies: number;
+}
+
+export interface BrawlerListResponse {
+	items: Brawler[];
+}