+++ packages/brawl-stars-api/src/lib/types/powers.ts
@@ -1,0 +1,10 @@
+interface Power {
+	id: number;
+	name: string;
+}
+
+export type StarPower = Power;
+export type Gadget = Power;
+export interface Gear extends Power {
+	level: number;
+}