+++ packages/brawl-stars-api/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @brawltracker/brawl-stars-api
 
+## 1.1.3
+
+### Patch Changes
+
+- - Add more routes to brawl-stars-api
+  - Fix some issues in api-utils
+- Updated dependencies
+  - @brawltracker/supercell-api-utils@1.1.3
+
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @brawltracker/supercell-api-utils@1.1.2
+
 ## 1.1.1
 
 ### Patch Changes