+++ packages/supercell-api-utils/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @brawltracker/supercell-api-utils
 
+## 1.1.3
+
+### Patch Changes
+
+- - Add more routes to brawl-stars-api
+  - Fix some issues in api-utils
+
+## 1.1.2
+
+### Patch Changes
+
+- fix files in npm
+
 ## 1.1.1
 
 ### Patch Changes